comments

+++ tests/end_to_end/sh-negc/manual-out.c
@@ -1,15 +1,15 @@
 u32 test(u32 value) {
-    return (u32) -(s32) value;
+    return -value;
 }
 
 u32 test_set(u32 value) {
-    return -(s32) value - 1;
+    return -value - 1;
 }
 
 u32 test_chain(u32 low, u32 high) {
-    return -(s32) high - (low != 0);
+    return -high - (low != 0);
 }
 
 u32 test_set_chain(u32 low, u32 high) {
-    return -(s32) high - 1;
+    return -high - 1;
 }
